Adding React Redux and code cleaning

+++ server.js
@@ -90,31 +90,35 @@ app.post('/update_profile', function(request, response){
 app.post('/create_project', function(req, res){
   let form = new multiparty.Form();
   form.parse(req, (err, fields, files) => {
-  let { path: tempPath, originalFilename } = files.file[0];
-  var fileName = + new Date() + originalFilename.replace(/\s/g, '');
+    if(files.file != undefined){
+      let { path: tempPath, originalFilename } = files.file[0];
+      var fileName = + new Date() + originalFilename.replace(/\s/g, '');
+      
+      let copyToPath = "./src/project-file/" + fileName; 
+      //add path (copyToPath) to database pending 
+      console.log(copyToPath);
+      fs.readFile(tempPath, (err, data) => {
+      if (err) throw err;
+      fs.writeFile(copyToPath, data, (err) => {
+      if (err) throw err;
+      // delete temp image
+      console.log(fields);
+      fs.unlink(tempPath, () => {
+      });
+      
+      });
+      });
+    }
   
-  let copyToPath = "./src/project-file/" + fileName; 
-  //add path (copyToPath) to database pending 
-  console.log(copyToPath);
-  fs.readFile(tempPath, (err, data) => {
-  if (err) throw err;
-  fs.writeFile(copyToPath, data, (err) => {
-  if (err) throw err;
-  // delete temp image
-  console.log(fields);
-  fs.unlink(tempPath, () => {
-  });
-  var sql= "INSERT into Project(title, description, skills_required, min_budget, max_budget, user_id, created_at, file_name) values ('" + fields.title[0] + "',  '" + 
-  fields.description[0]+ "', '" + fields.skills_required[0] + "', '" +  fields.minimum_budget[0] + "', '" + fields.maximum_budget[0] + "', '" 
-  + fields.user_id[0] + "', '" + new Date().toLocaleString() + "', '" + fileName + "')";
-  console.log(sql);
-  con.query(sql,function(err,rows){
-    if(err) throw err ;
-  });
-  });
-  });
+    var sql= "INSERT into Project(title, description, skills_required, min_budget, max_budget, user_id, created_at, file_name) values ('" + fields.title[0] + "',  '" + 
+    fields.description[0]+ "', '" + fields.skills_required[0] + "', '" +  fields.minimum_budget[0] + "', '" + fields.maximum_budget[0] + "', '" 
+    + fields.user_id[0] + "', '" + new Date().toLocaleString() + "', '" + fileName + "')";
+    console.log(sql);
+    con.query(sql,function(err,rows){
+      if(err) throw err ;
+    });
 });
-  res.json({message: "hello"});
+  res.json({message: "Project Created"});
 });
 
 app.get('/get_all_projects', function(request, response){
@@ -169,6 +173,11 @@ app.post('/submit_bid', function(request, response){
     }
     con.query(sql_query,function(err,rows){
       if(err) throw err ;
+      //response.json({rows: rows})
+    });
+    var get_total_bid = "Select AVG(number_of_days) as avgDays from Bid where project_id='" + request.body.project_id + "' Group By project_id";
+    con.query(get_total_bid,function(err,rows){
+      if(err) throw err ;
       response.json({rows: rows})
     });
   });
@@ -189,27 +198,39 @@ app.post('/hire_user', function(request, response){
   var query = "Update Bid SET status = 'Accepted' where user_id= '" + request.body.free_lancer_id + "' and project_id= '" + 
   request.body.p_id + "'";
   
-  var query1 = "UPDATE Project SET assigned_to = '" + request.body.free_lancer_id + "' where id='" + request.body.p_id + "'";
+  //var query1 = "UPDATE Project SET assigned_to = '" + request.body.free_lancer_id + "' where id='" + request.body.p_id + "'";
   var query2 = "UPDATE Bid Set status = 'Rejected' where project_id='" + request.body.p_id + "' and user_id != '" + request.body.free_lancer_id + "'" ;
+  var query3 = "Select * from Bid where project_id='" + request.body.p_id + "' and status='Accepted'";
 
-  con.query(query1, function(err,rows){
+  con.query(query, function(err,rows){
     if(err) throw err;
-    console.log(rows.length);
     //rows.length >= 1 ? response.json({dataUpdated: true, rows: rows[0]}) :  response.json({dataUpdated: false});
-  })
+  });
 
   con.query(query2, function(err,rows){
     if(err) throw err;
     console.log(rows.length);
-    //rows.length >= 1 ? response.json({dataUpdated: true, rows: rows[0]}) :  response.json({dataUpdated: false});
   })
 
-
-  con.query(query, function(err,rows){
+  con.query(query3, function(err,rows){
     if(err) throw err;
-    console.log(rows.length);
-    rows.length >= 1 ? response.json({dataUpdated: true, rows: rows[0]}) :  response.json({dataUpdated: false});
-  });
+    console.log(rows);
+
+
+    if(rows.length >= 1){
+      var query4 = "UPDATE Project Set assigned_to='" + request.body.free_lancer_id + 
+        "', date_of_completion= DATE_ADD(NOW() , INTERVAL " + rows[0].number_of_days + " DAY) where id='" + request.body.p_id + "'"; 
+        con.query(query4, function(err, rows){
+          if(err) throw err;
+          rows.length >= 1 ? response.json({dataUpdated: true, rows: rows[0]}) :  response.json({dataUpdated: false});
+          
+      });
+    }
+    
+    
+  })
+
+  
 });
 
 app.get('/get_all_user_bid_projects', function(request, response){
@@ -225,11 +246,12 @@ app.get('/get_all_user_bid_projects', function(request, response){
 });
 
 app.get('/get_all_user_published_projects', function(request, response){
+  console.log(request.query)
   var sql = "select b.id,b.title, b.assigned_to as freelancer_id, b.user_id as employer_id, b.date_of_completion as date_of_completion, avgTable.avgDays,C.name as owner ,b.assigned_to as freeLancer, " +
   "CASE WHEN b.assigned_to is null or trim(b.assigned_to) = '' " +
       "Then '' ELSE  (select D.name from User D where D.id=b.assigned_to) END as freelancer_name from " +
   "(SELECT b.id,COALESCE(avg(a.number_of_days),0) as avgDays from Project b left OUTER JOIN Bid a on a.project_id=b.id group by b.id) as avgTable " +
-  ",Project b, User C where b.user_id = '125' and b.user_id=C.id and avgTable.id=b.id group by b.id, b.title,avgTable.avgDays,owner,freeLancer"
+  ",Project b, User C where b.user_id = '" + request.query.u_id + "' and b.user_id=C.id and avgTable.id=b.id group by b.id, b.title,avgTable.avgDays,owner,freeLancer"
 
   con.query(sql,function(err,rows){
     if(err) throw err;
@@ -242,6 +264,15 @@ app.post('/get-bid-value-for-user', function(request, response){
   var sql = "Select * from Bid where user_id = '" + request.body.user_id + "' and project_id = '" + request.body.project_id + "'";
   console.log(sql)
   con.query(sql,function(err,rows){
+    if(err) throw err;
+    console.log(rows);
+    rows.length >= 1 ? response.json({data_present: true, rows: rows[0]}) :  response.json({data_present: false});
+  });
+});
+
+app.get('/get-user-name', function(request, response){
+  var sql = "select * from User where id='" + request.query.id + "'";
+  con.query(sql, function(err, rows){
     if(err) throw err;
     console.log(rows);
     rows.length >= 1 ? response.json({data_present: true, rows: rows[0]}) :  response.json({data_present: false});
@@ -279,62 +310,3 @@ app.post('/upload-Image', function(req, res){
 app.listen(port, function() {
  console.log(`api running on port ${port}`);
 });
-
-
-
-
-
-
-
-// SELECT Project.*, User.name, count(Bid.project_id) as total_bids from Project left join Bid on (Project.id = Bid.project_id) LEFT JOIN User on 
-// (Project.user_id = User.id) where Project.user_id = '100' group by Project.id
-// SELECT Project.*, User.name, AVG(Bid.number_of_days) as total_bids from Project left join Bid on (Project.id = Bid.project_id) LEFT JOIN User on 
-// (Project.user_id = User.id) where Project.user_id = '100' group by Project.id
-
-// Select Project.*, User.name, Bid.number_of_days, Project.id, AVG(DISTINCT Bid.number_of_days) as days from Project left join Bid on (Project.id = Bid.project_id) LEFT JOIN User on (Project.user_id = User.id) where Bid.user_id = '100' GROUP BY Bid.id
-
-// Select AVG(Bid.number_of_days) as total, Project.title, User.name from Bid left JOIN Project on (Bid.project_id = Project.id) LEFT JOIN User on (Bid.user_id = User.id) GROUP BY Bid.id
-
-
-// SELECT AVG(Bid.number_of_days), Project.id, Project.title from Bid INNER JOIN Project on (Bid.project_id=Project.id) GROUP BY Project.id
-// SELECT AVG(Bid.number_of_days), Bid.number_of_days, Project.id, Project.title from Bid INNER JOIN Project on (Bid.project_id=Project.id) where Bid.user_id = '100' GROUP BY Project.id
-// SELECT AVG(Bid.number_of_days), User.name, Bid.number_of_days, Project.id, Project.title from Bid INNER JOIN Project on (Bid.project_id=Project.id) INNER JOIN User on (Bid.user_id = User.id) where Bid.user_id = '100' GROUP BY Project.id, Bid.number_of_days
-
-
-
-
-
-// select averageTable.avgDays,title,averageTable.id,X.number_of_days,c.name, c.id
-// from (select avg(b.number_of_days) as avgDays,p.title,p.id from Bid b,Project p where b.project_id=p.id group by p.id) as averageTable, Bid X ,User c where X.project_id=averageTable.id and X.user_id=100 
-// and X.user_id =c.id
-
-
-// select b.id,b.title,avgTable.avgP,C.name as owner ,b.assigned_to as freeLancer from (SELECT b.id,COALESCE(avg(a.price),0) as avgP from Project b left OUTER JOIN Bid a on a.project_id=b.id group by b.id) as avgTable ,Project b, User C where b.user_id = '125' and b.user_id=C.id and avgTable.id=b.id group by b.id, b.title,avgTable.avgP,owner,freeLancer
-
-// select b.id,b.title,avgTable.avgP,C.name as owner ,b.assigned_to as freeLancer,
-// CASE
-// 	WHEN b.assigned_to is null or trim(b.assigned_to) = ""
-//     Then ""
-//     ELSE
-// 	(select D.name from User D where D.id=b.assigned_to) 
-//     END
-
-// from 
-// (SELECT b.id,COALESCE(avg(a.price),0) as avgP from Project b left OUTER JOIN Bid a on a.project_id=b.id group by b.id) as avgTable
-// ,Project b, User C where b.user_id = '125' and b.user_id=C.id and avgTable.id=b.id 
-
-// group by b.id, b.title,avgTable.avgP,owner,freeLancer
-
-// select b.id,b.title,avgTable.avgP,C.name as owner ,b.assigned_to as freeLancer,
-// CASE
-// 	WHEN b.assigned_to is null or trim(b.assigned_to) = ""
-//     Then ""
-//     ELSE
-// 	(select D.name from User D where D.id=b.assigned_to) 
-//     END
-
-// from 
-// (SELECT b.id,COALESCE(avg(a.price),0) as avgP from Project b left OUTER JOIN Bid a on a.project_id=b.id group by b.id) as avgTable
-// ,Project b, User C where b.user_id = '125' and b.user_id=C.id and avgTable.id=b.id 
-
-// group by b.id, b.title,avgTable.avgP,owner,freeLancer
